refactor: extract fuzzy search and section builder from SpotlightSearch
Move the fuzzy scoring algorithm to fuzzyScore.ts and the section
building logic to spotlightSections.ts, reducing SpotlightSearch.tsx
from 430 lines to ~170 lines of focused UI code.
